Fix hunt position value
Fixes the position to be "" instead of "()" when the position is not
reported, resolving an issue where dead/missing hunts were incorrectly
marked as alive. Tweaks formatting of the callout to put the coordinates
next to the zone name.

Now that instances have been removed, the bot works for 5.x zones.